feat: add PeopleList.js
