+++ src/components/PeopleList/index.js
@@ -1,0 +1,3 @@
+import PeopleList from './PeopleList';
+
+export default PeopleList;
